requirements.txt
requests
schedule
paho-mqtt
https://github.com/beyond-code-github/pyenergenie/archive/master.zip#pyenergenie==0.3.0